requirements.txt
flask
google-generativeai
loguru
requests
faiss-cpu
python-dotenv
pyyaml
sentence-transformers
